SINUS + tinyxml. NOTE: PARSER NE MARCHE PAS.
