+++ src/src/Phenomene.cpp
@@ -11,7 +11,6 @@
 #include <string>
 #include <cstdlib>
 #include <ctime>
-#include <math.h>
 
 namespace Domotique {
 using namespace std;
