+++ src/src/Phenomene_sinus.cpp
@@ -9,13 +9,15 @@
 #include <iostream>
 
 using namespace std;
+using namespace math_const;
 namespace Domotique {
 
 Phenomene_sinus::Phenomene_sinus(string nom, double ampl, long int period,
 		double sat_max,
 		double sat_min,
-		double offs, long int phase): Phenomene(nom), ampl_(ampl), period_(period),
-		sat_max_(sat_max), sat_min_(sat_min), offs_(offs), phase_(phase){
+		double offs, long int phase): Phenomene(nom),tic_(0),  ampl_(ampl),
+									period_(period), sat_max_(sat_max),
+									sat_min_(sat_min), offs_(offs), phase_(phase){
 	cout<< "[Phenomene_sinus] Création de "<< nom << endl;
 	assert(period >0);
 }
@@ -25,7 +27,11 @@ Phenomene_sinus::~Phenomene_sinus() {
 }
 void Phenomene_sinus::run(void)
 {
-	// TODO Run
+	double offs = box_mull(offs_);
+	double ampl = box_mull(ampl_);
+	double y = offs+ ampl*sin(2*PI*(tic_+phase_)/period_);
+	p_etat_->put_valphen(y);
+	tic_++;
 }
 
 } /* namespace Domotique */
