+++ src/src/tinyxml_epfl/tinyxml.h
@@ -21,8 +21,6 @@ must not be misrepresented as being the original software.
 3. This notice may not be removed or altered from any source
 distribution.
 */
-
-
 #ifndef TINYXML_INCLUDED
 #define TINYXML_INCLUDED
